Move classes as suggested; provide missing descriptions
