+++ src/main/resources/META-INF/rewrite/slf4j.yml
@@ -186,8 +186,8 @@ tags:
   - java-util-logging
   - slf4j
 recipeList:
-  - org.openrewrite.java.logging.jul.GetLoggerToLoggerFactoryRecipes
-  - org.openrewrite.java.logging.jul.LoggerIsLoggableRecipes
+  - org.openrewrite.java.logging.slf4j.JulGetLoggerToLoggerFactoryRecipes
+  - org.openrewrite.java.logging.slf4j.JulIsLoggableToIsEnabledRecipes
   - org.openrewrite.staticanalysis.SimplifyConstantIfBranchExecution
   - org.openrewrite.java.logging.log4j.JulToLog4j
   - org.openrewrite.java.logging.slf4j.Log4j2ToSlf4j1
